WIP: fixing rate notifications

+++ src/db/schema/userNotifications.ts
@@ -1,4 +1,4 @@
-import { integer, pgTable, varchar } from 'drizzle-orm/pg-core'
+import { integer, numeric, pgTable, varchar } from 'drizzle-orm/pg-core'
 import { tokens, wallets } from '.'
 
 export const userNotifications = pgTable('user_notifications', {
@@ -9,5 +9,5 @@ export const userNotifications = pgTable('user_notifications', {
     .notNull()
     .references(() => tokens.token),
   timestamp: integer('timestamp').notNull(),
-  price: integer('price').notNull(),
+  price: numeric('price').notNull(),
 })
